Fixed: AutoCompleteSolrFacet API fails due to invalid characters when using GET method (#2gagqqt) Error: Invalid character found in the request target [&#47;api&#47;AutoCompleteSolrFacet?facetToSelect=tagsFacet&docType=PRODUCT&coreName=enterpriseSearch&searchfield=tags&jsonQuery=encodeURIComponent({%22query%22:%22*:*%22,%22filter%22:%20[%22docType:PRODUCT%22]})&noConditionFind=N&q=&term=&offset=&limit=%2210%22 ]. The valid characters are defined in RFC 7230 and RFC 3986
Issue:
If you use an upper version of Tomcat 8.5 it throws this exception if the URL path contains '[' and ']'. For older versions, it works.

Solution:
Used paramsSerializer available with axios to serialize params and qs npm package (https://www.npmjs.com/package/qs) for stringifying the parameters.
When passing object to qs, bracket or dot notation is returned as output. Handled the case by stringify the object before passing to qs.stringify() method